nganiriza_backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
diskcache==5.6.3
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llama_cpp_python==0.3.16
Markdown==3.10
MarkupSafe==3.0.3
numpy==1.26.4
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.28.0
gunicorn==22.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
